Pull ConvertToRasterUsingResourceContext into its own header file and tweak the test
